Update statistics event to use long as statistics do

+++ monolith-api/src/main/scala/gg/warcraft/monolith/api/player/events.scala
@@ -138,7 +138,7 @@ case class PlayerRespawnEvent(
 case class PlayerStatisticsChangedEvent(
     player: Player,
     statistic: String,
-    amount: Int,
-    newCurrent: Int
+    amount: Long,
+    newCurrent: Long
 ) extends Event
     with PlayerEvent
